Create GB7MBB-2mb-link-upgrade
This is a document I updated back in 1997 which I used to plan and keep notes on a project I was taking when I worked at Lancaster University. Essentially I am creating a faster, 2Mb link, from the University to my house. This is the year that the BBC intriduced their own website.